requirements.txt
cherrypy
simpletalsix >= 6.2.0.dev0
